feat: add Header and Footer components, and move Nav component to _shared folder
- create Header component
- create Footer component
- move Nav component to _shared folder
