+++ Front/FinanceTracker.Client/src/app/_shared/footer/footer.component.html
@@ -1,0 +1,1 @@
+<p>footer works!</p>
