+++ Front/FinanceTracker.Client/src/app/_shared/header/header.component.html
@@ -1,0 +1,1 @@
+<p>header works!</p>
