+++ Front/FinanceTracker.Client/src/app/dashboard/dashboard.component.html
@@ -1,0 +1,1 @@
+<p>dashboard works!</p>
